Rescaled so its not proportional to the real life stylus
Took the dimensions of the real life stylus and scaled them up by 10 percent  (ie multipled the deminsions by 10) so that the new model worked with the buttons and rotaing ball that Logan had for the original stylus. However, and this will come with testing, the stylus may still be on the small side. If that's the case, I will need to discus with logan about sizing all elements of the stylus so that they work and are proportional.